Refactor qbvh into multiple files
